Replace QtKeychain with a submodule
This commit matches up with the date on which the source tree was
added (Before 31 August 2021) and the most recent commit's content
is already present, so I'm fairly confident this is the exact
revision.

This allows easier updating if required, plus it preserves history.

Added a note to COMPILATION.md about cloning with submodules.